cleaned up the code
Score is now a part of the HUD. Ball now declares and adds an impact
emitter to the world it's in, rather than the world declaring the impact
emitter and assigning it to the ball. Ball now passes a reference of
itself to the impact emitter for the functions that need it. This format
of an object passing a reference of itself to the impact emitter will
allow me to write other functions in impact emitter that other entities
can use.